fix: juju credential discovery.
A test found that we were sending both the client ID/secret AND the
username/password when only one is allowed. Refactored the credential
discovery logic to split out env var values vs live discovery values.
Client ID/secret must be specified in the plan and/or environment
variables, if we've looked at that data and still do not have full
authentication data, fail here.

If live discovery is possible, it will always find the username/password
values, and never the client ID/secret values.