fix: errors seen in TestAcc_ResourceApplication_Updates
The storge attribute of the schema should not be using
UseStateForUnknown as computed only attributes. This confuses terraform
which expects the plan to match what is put into state. Storage objects
won't change, but more can be added which is the problem.

Add a expected error for step 2 as it's around placement. An github
issue has been created where the problem was seen in the wild. Placement
directives and actual machine values do not match.